backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
pinecone-client==5.0.1
python-dotenv==1.0.1
pydantic-settings==2.7.0
sentence-transformers==3.3.1
